update_utils/requirements.txt
requests
tqdm
sentence-transformers
numpy
faiss-cpu
fastapi 
uvicorn[standard]
PyJWT
python-multipart
Pillow
opencv-python
ultralytics
pytesseract
openai
replicate
psycopg2-binary
authlib

